Fix ddl_script.sql

+++ docs/ddl_script.sql
@@ -17,7 +17,7 @@ CREATE TABLE `cryptoCurrency` (
 CREATE TABLE `cryptoWallet` (
     `idWallet` INTEGER not NULL AUTO_INCREMENT,
     `idCrypto` INTEGER NOT NULL,
-    `purchasePrice` double(20,19) not NULL,
+    `purchasePrice` double(30,9) not NULL,
     `quantity` double(20,19) NOT NULL,
     `purchaseDate` DATETIME NOT NULL default CURRENT_TIMESTAMP,
     PRIMARY KEY (`idWallet`),
